refactor: create filter to prisma error
