+++ api/src/errors/custom-error.ts
@@ -1,14 +1,10 @@
 class CustomError extends Error {
   status: number;
+
   code: string | undefined;
-  constructor(options: {
-    code?: string;
-    message: string;
-    status: number;
-  }) {
-    super(
-      options.message
-    )
+
+  constructor(options: { code?: string; message: string; status: number }) {
+    super(options.message);
     this.code = options.code;
     this.status = options.status;
   }
