+++ api/test/update-user.spec.ts
@@ -1,4 +1,4 @@
-import userServices from '../src/aaa/users.service';
+import userServices from '../src/services/users.service';
 
 describe('User Services', () => {
   it('should update a user in the database', async () => {
